+++ tests/Toarnbeike.Unions.SourceGenerator.Tests/SourceGenerated/StatusUnion/Status.cs
@@ -1,10 +1,8 @@
-﻿using Toarnbeike.Unions.SourceGenerated.Complex;
-using Toarnbeike.Unions.SourceGenerator;
+﻿using Toarnbeike.Unions.SourceGenerator;
 
 namespace Toarnbeike.Unions.SourceGenerated.StatusUnion;
-
 
 [UnionCase(typeof(Active))]
 [UnionCase(typeof(Retry))]
 [UnionCase(typeof(Aborted))]
-public partial class Status { }
+public partial class Status;